Simplify / optimize complex card
Bindable layouts in an item template are rather heavy.  While this change does not allow for unlimited/dynamic reviews or types of social media buttons, I think it's safe to say that the amount expected in each item for both cases is finite.  I added some properties to the data model and duplicated the item templates for the bindable layouts so there's a prefixed amount, which is going to be cheaper to hide extra controls so we aren't re-creating all of the views for each cell in the bindable layouts each time the bindingcontext on the recycled cell changes.  This of course can be changed to accommodate the max number of reviews and social media buttons ever expected to highlight per item.