Use a more dynamic approach but without bindable layout still
This caches the views that are being dynamically added/removed from the item template's layout when the item view's bindingcontext changes.

The difference here from a bindable layout is that instead of recreating all the child views every time the binding context changes, we are keeping instances of the child views cached, so the recycled cell will add more child views as needed (first from the cache, or creates new if there aren't any cached ones available), or remove child views if the binding context change requires fewer to be displayed, but then it caches those for future use.